Move each declarator extractor on its own file but stil in the same pkg
